chore: Update FreeRADIUS configuration and Accounting setup
- Fix FreeRADIUS Docker setup (custom build with libpq in Dockerfile)
- Correct Docker volume mounts for Alpine image layout (/opt/etc/raddb)
- Configure SQL module for PostgreSQL backend (read_clients=yes, custom client_query)
- Enable Multi-NAS support via database 'nas' table
- Fix Database Schema: Add missing columns to 'radacct' (IPv6, acctupdatetime) for correct accounting
- Clean up install.sh: Auto-configure FreeRADIUS and default NAS entries
- Remove temporary verification files

+++ backend/routes/api/v1/broadcast.js
@@ -12,7 +12,7 @@ router.get('/', async (req, res) => {
         bm.*,
         u.username as created_by_name
       FROM broadcast_messages bm
-      LEFT JOIN users u ON bm.created_by = u.id
+      LEFT JOIN users u ON bm.created_by::integer = u.id
       ORDER BY bm.priority ASC, bm.created_at DESC
     `;
 
@@ -57,7 +57,7 @@ router.get('/messages', async (req, res) => {
         bm.*,
         u.username as created_by_name
       FROM broadcast_messages bm
-      LEFT JOIN users u ON bm.created_by = u.id
+      LEFT JOIN users u ON bm.created_by::integer = u.id
       ORDER BY bm.created_at DESC
     `;
